# firewatch-frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "BulkReassignRequest": {
        "properties": {
          "owner_id": {
            "title": "Owner Id",
            "type": "integer"
          },
          "risk_ids": {
            "items": {
              "type": "string"
            },
            "maxItems": 200,
            "minItems": 1,
            "title": "Risk Ids",
            "type": "array"
          }
        },
        "required": [
          "risk_ids",
          "owner_id"
        ],
        "title": "BulkReassignRequest",
        "type": "object"
      },
...
      "BulkRiskResult": {
        "properties": {
          "errors": {
            "items": {
              "$ref": "#/components/schemas/BulkRiskError"
            },
            "title": "Errors",
            "type": "array"
          },
          "updated": {
            "items": {
              "type": "string"
            },
            "title": "Updated",
            "type": "array"
          }
        },
        "required": [
          "updated",
          "errors"
        ],
        "title": "BulkRiskResult",
        "type": "object"
      },
      "BulkStatusRequest": {
        "properties": {
          "risk_ids": {
            "items": {
              "type": "string"
            },
            "maxItems": 200,
            "minItems": 1,
            "title": "Risk Ids",
            "type": "array"
          },
          "status": {
            "$ref": "#/components/schemas/RiskStatus"
          }
        },
        "required": [
          "risk_ids",
          "status"
        ],
        "title": "BulkStatusRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RiskListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RiskResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "items"
        ],
        "title": "RiskListResponse",
        "type": "object"
      },
...
      "RiskStatus": {
        "enum": [
          "open",
          "in_progress",
          "mitigated",
          "accepted",
          "closed"
        ],
        "title": "RiskStatus",
        "type": "string"
      },
...
    }
  },
  "info": {
    "description": "Cybersecurity Risk Register API -- NIST 800-30 aligned",
    "title": "Firewatch",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/risks": {
      "get": {
        "operationId": "list_risks_api_risks_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/RiskStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "in": "query",
            "name": "owner_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Owner Id"
            }
          },
          {
            "in": "query",
            "name": "due_for_review",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Due For Review"
            }
          },
          {
            "in": "query",
            "name": "skip",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Skip",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "cookie",
            "name": "access_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Risks",
        "tags": [
          "Risks"
        ]
      },
      "post": {
        "operationId": "create_risk_api_risks_post",
        "parameters": [
          {
            "in": "cookie",
            "name": "access_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RiskCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Risk",
        "tags": [
          "Risks"
        ]
      }
    },
    "/api/risks/bulk/reassign": {
      "post": {
        "description": "Reassign owner_id on multiple risks. Admin/analyst only.",
        "operationId": "bulk_reassign_api_risks_bulk_reassign_post",
        "parameters": [
          {
            "in": "cookie",
            "name": "access_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkReassignRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkRiskResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Bulk Reassign",
        "tags": [
          "Risks"
        ]
      }
    },
    "/api/risks/bulk/rescore": {
      "post": {
        "description": "Append a new assessment (likelihood/impact) to multiple risks.",
        "operationId": "bulk_rescore_api_risks_bulk_rescore_post",
        "parameters": [
          {
            "in": "cookie",
            "name": "access_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkRescoreRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkRiskResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Bulk Rescore",
        "tags": [
          "Risks"
        ]
      }
    },
    "/api/risks/bulk/status": {
      "post": {
        "description": "Change status on multiple risks. Per-risk ownership enforced by the service.",
        "operationId": "bulk_status_api_risks_bulk_status_post",
        "parameters": [
          {
            "in": "cookie",
            "name": "access_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkStatusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkRiskResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Bulk Status",
        "tags": [
          "Risks"
        ]
      }
    },
    "/api/risks/export": {
      "get": {
        "description": "Download all risks visible to the current user as a CSV file.",
        "operationId": "export_risks_api_risks_export_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "access_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Risks",
        "tags": [
          "Risks"
        ]
      }
    },
    "/api/risks/import": {
      "post": {
        "description": "Bulk-create risks from an uploaded CSV. Errors are reported per row.",
        "operationId": "import_risks_api_risks_import_post",
        "parameters": [
          {
            "in": "cookie",
            "name": "access_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_risks_api_risks_import_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Import Risks",
        "tags": [
          "Risks"
        ]
      }
    },
    "/api/risks/import-template": {
      "get": {
        "description": "Download a CSV template (header + one example row) for risk import.",
        "operationId": "import_template_api_risks_import_template_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "access_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Import Template",
        "tags": [
          "Risks"
        ]
      }
    },
    "/api/risks/{risk_id}": {
      "delete": {
        "description": "Soft-delete a risk. Admin only. The record is retained for audit purposes.",
        "operationId": "delete_risk_api_risks__risk_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "risk_id",
            "required": true,
            "schema": {
              "title": "Risk Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "access_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Risk",
        "tags": [
          "Risks"
        ]
      },
      "get": {
        "operationId": "get_risk_api_risks__risk_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "risk_id",
            "required": true,
            "schema": {
              "title": "Risk Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "access_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Risk",
        "tags": [
          "Risks"
        ]
      },
      "put": {
        "operationId": "update_risk_api_risks__risk_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "risk_id",
            "required": true,
            "schema": {
              "title": "Risk Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "access_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RiskUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Risk",
        "tags": [
          "Risks"
        ]
      }
    },
...
  }
}